remove @objc SVGImage
