+++ SwiftDraw/SVG.swift
@@ -34,13 +34,55 @@ import Foundation
 #if canImport(CoreGraphics)
 import CoreGraphics
 
-@objc(SVGImage)
-public final class SVG: NSObject {
+public struct SVG {
     public let size: CGSize
 
     //An Image is simply an array of CoreGraphics draw commands
     //see: Renderer.swift
     let commands: [RendererCommand<CGTypes>]
+
+    public init?(fileURL url: URL, options: SVG.Options = .default) {
+        do {
+            let svg = try DOM.SVG.parse(fileURL: url)
+            self.init(dom: svg, options: options)
+        } catch {
+            XMLParser.logParsingError(for: error, filename: url.lastPathComponent, parsing: nil)
+            return nil
+        }
+    }
+
+    public init?(named name: String, in bundle: Bundle = Bundle.main, options: SVG.Options = .default) {
+        guard let url = bundle.url(forResource: name, withExtension: nil) else {
+            return nil
+        }
+
+        self.init(fileURL: url, options: options)
+    }
+
+    public init?(data: Data, options: SVG.Options = .default) {
+        guard let svg = try? DOM.SVG.parse(data: data) else {
+            return nil
+        }
+
+        self.init(dom: svg, options: options)
+    }
+
+    public struct Options: OptionSet {
+        public let rawValue: Int
+        public init(rawValue: Int) {
+            self.rawValue = rawValue
+        }
+
+        public static let hideUnsupportedFilters = Options(rawValue: 1 << 0)
+
+        public static let `default`: Options = []
+    }
+}
+
+@available(*, unavailable, renamed: "SVG")
+public enum Image { }
+
+extension SVG {
 
     init(dom: DOM.SVG, options: Options) {
         self.size = CGSize(width: dom.width, height: dom.height)
@@ -60,25 +102,11 @@ public final class SVG: NSObject {
             generator.renderCommands(for: layer)
         )
     }
-
-    public struct Options: OptionSet {
-        public let rawValue: Int
-        public init(rawValue: Int) {
-            self.rawValue = rawValue
-        }
-
-        public static let hideUnsupportedFilters = Options(rawValue: 1 << 0)
-
-        public static let `default`: Options = []
-    }
 }
-
-@available(*, unavailable, renamed: "SVG")
-public enum Image { }
 
 #else
 
-public final class SVG: NSObject {
+public struct SVG {
     public let size: CGSize
 
     init(dom: DOM.SVG, options: Options) {
@@ -116,69 +144,3 @@ public extension SVG {
     }
 }
 #endif
-
-extension DOM.SVG {
-
-    static func parse(fileURL url: URL, options: XMLParser.Options = .skipInvalidElements) throws -> DOM.SVG {
-        let element = try XML.SAXParser.parse(contentsOf: url)
-        let parser = XMLParser(options: options, filename: url.lastPathComponent)
-        return try parser.parseSVG(element)
-    }
-
-    static func parse(data: Data, options: XMLParser.Options = .skipInvalidElements) throws -> DOM.SVG {
-        let element = try XML.SAXParser.parse(data: data)
-        let parser = XMLParser(options: options)
-        return try parser.parseSVG(element)
-    }
-}
-
-public extension SVG {
-
-    convenience init?(fileURL url: URL, options: SVG.Options = .default) {
-        do {
-            let svg = try DOM.SVG.parse(fileURL: url)
-            self.init(dom: svg, options: options)
-        } catch {
-            XMLParser.logParsingError(for: error, filename: url.lastPathComponent, parsing: nil)
-            return nil
-        }
-    }
-
-    convenience init?(named name: String, in bundle: Bundle = Bundle.main, options: SVG.Options = .default) {
-        guard let url = bundle.url(forResource: name, withExtension: nil) else {
-            return nil
-        }
-
-        self.init(fileURL: url, options: options)
-    }
-
-    convenience init?(data: Data, options: SVG.Options = .default) {
-        guard let svg = try? DOM.SVG.parse(data: data) else {
-            return nil
-        }
-
-        self.init(dom: svg, options: options)
-    }
-
-
-    struct Insets: Equatable {
-        public var top: CGFloat
-        public var left: CGFloat
-        public var bottom: CGFloat
-        public var right: CGFloat
-
-        public init(
-            top: CGFloat = 0,
-            left: CGFloat = 0, 
-            bottom: CGFloat = 0,
-            right: CGFloat = 0
-        ) {
-            self.top = top
-            self.left = left
-            self.bottom = bottom
-            self.right = right
-        }
-
-        public static let zero = Insets(top: 0, left: 0, bottom: 0, right: 0)
-    }
-}
